update of Entity "Country "  simplified in CoreData, type of tempMax and tempMin attribute changed on Int64, Page Control added, weekDay in DailyWeatherTableView displayed fully, backToMenu Button added

+++ WeatherApp/WeatherApp/Model/DailyWeather+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  DailyWeather+CoreDataProperties.swift
 //  WeatherApp
 //
-//  Created by Лера on 10/26/19.
+//  Created by Лера on 10/27/19.
 //  Copyright © 2019 com.vkravets. All rights reserved.
 //
 //
@@ -20,8 +20,8 @@ extension DailyWeather {
     @NSManaged public var dailyIcon: String?
     @NSManaged public var dailyTemp: Double
     @NSManaged public var date: Int64
-    @NSManaged public var tempMax: Double
-    @NSManaged public var tempMin: Double
+    @NSManaged public var tempMax: Int64
+    @NSManaged public var tempMin: Int64
     @NSManaged public var city: City?
 
 }
